+++ tests/python/pyexample/conf.py
@@ -19,3 +19,4 @@ autoapi_type = "python"
 autoapi_dirs = ["example"]
 autoapi_file_pattern = "*.py"
 autoapi_python_class_content = "both"
+autoapi_keep_files = True